more refactor, rename index.ts to version.ts
